Safer value handling when the input data is deleted
In my setup, downshift returns "undefined" when I delete all text in the input box of the Combobox, SelectMenu components. I am not sure why for you it returns an empty string, but I tried several combinations and it was always undefined. When it was undefined it led to an  exception. This should fix that.